move orchestrator to a separate file and use env for rpc url
